==> Jupyter Notebooks/FULL PIPELINE/0. Task Description.ipynb <==
# Multi-class classification of sentiment associated with therapies in English tweets

## Daniel Jimenez Campos - 2023

### Data Description

**Training data**: 3009 tweets  
**Validation data**: 753 tweets  
**Testing data**: TBA  
**Evaluation metric**: micro-averaged F1-score  

### Data Examples

| tweet_id | therapy     | text                                                                                                                                                       | label    |
|----------|-------------|------------------------------------------------------------------------------------------------------------------------------------------------------------|----------|
| 15309    | meditation  | Did you know meditation can be one of the *most rewarding important things you do in your life*? Did you also know it’s *impossible to not be able to meditate*? For people that believe your mind must somehow go blank you’re wrong unless you’re dead. | positive |
| 15262    | acupuncture | abt to get acupuncture for my migraines for the first time ever & i am *terrified*                                                                             | neutral  |

### Submission Format

Please use the format below for submission. Submissions should contain tweet_id and label separated by tabspace in the same order as below.

tweet_id label
15309 positive
15262 neutral


The unzipped submission data needs to be named as *"answer.txt"* and be zipped.

For more information, please refer to [this link](https://github.com/codalab/codalab-competitions/wiki/User_Building-a-Scoring-Program-for-a-Competition#directory-structure-for-submissions).


Simply copy and paste the above Markdown-formatted text into an empty Markdown cell in Jupyter Notebook to display it correctly.

# Pipeline Index

0. Introduction
1. Exploratory Data Analysis
2. Data Cleaning
3. Feature Engineering
4. Hyperparameter Tuning
4. Vectorizing
5. Model Evaluation
6. Final model selection
7. Make predictions on new data
