==> Jupyter Notebooks/FULL PIPELINE/5. Vectorizing .ipynb <==
# Vectorizing - TD-IDF

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer

data = pd.read_csv("C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\data_cleaned_features.csv")

test_data = pd.read_csv("C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\test_data_cleaned_features.csv")

test_data.head()

print(len(data))
print(len(test_data))

### Split into train/test

# Splitting the data into training and testing sets

# The 'label' column is used as the target variable (y)
# The rest of the columns except 'tweet_id' & 'therapy' are used as the features (X)
# The test_size parameter is set to 0.2, which means 20% of the data will be used for testing

X_train, X_test, y_train, y_test = train_test_split(data[
    ['tweet_id', 'therapy', 'cleaned_text', 
     'avg_word_length', 'sia_positive_word_rate', 'sia_negative_word_rate', 'neutral_score', 'stopword_count', 
     'body_len', 'compound_score', 'punct%', 'positive_score', 'negative_score', 'neutral_score']
], data['label'], test_size=0.2)

X_train = data[['tweet_id', 'therapy', 'cleaned_text',
               'avg_word_length', 'sia_positive_word_rate', 'sia_negative_word_rate',
               'neutral_score', 'stopword_count', 'body_len', 'compound_score',
               'punct%', 'positive_score', 'negative_score', 'neutral_score']]

X_test = test_data[['tweet_id', 'therapy', 'cleaned_text',
                    'avg_word_length', 'sia_positive_word_rate', 'sia_negative_word_rate',
                    'neutral_score', 'stopword_count', 'body_len', 'compound_score',
                    'punct%', 'positive_score', 'negative_score', 'neutral_score']]

y_train = data['label']
y_test = test_data['label']

### Write out data

X_train.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\train_features.csv', mode='w', index=False)
# X_val.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\val_features.csv', mode='w', index=False)
X_test.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\test_features.csv', mode='w', index=False)

y_train.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\train_labels.csv', mode='w', index=False)
# y_val.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\val_labels.csv', mode='w', index=False)
y_test.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\test_labels.csv', mode='w', index=False)

### TD-IDF Vectorizer

# Creating a TfidfVectorizer object with the analyzer parameter set to the clean_text function
tfidf_vect = TfidfVectorizer()

# Fitting the TfidfVectorizer on the 'text' column of the training set
tfidf_vect_fit = tfidf_vect.fit(X_train['cleaned_text'])

# Transforming the 'text' column of the training and testing sets into TF-IDF features
tfidf_train = tfidf_vect_fit.transform(X_train['cleaned_text'])
tfidf_test = tfidf_vect_fit.transform(X_test['cleaned_text'])

# Concatenating the features columns with the TF-IDF features of the training set
X_train_vect = pd.concat([X_train[
    ['tweet_id', 'avg_word_length', 'sia_positive_word_rate', 'sia_negative_word_rate', 'neutral_score', 
     'stopword_count', 'body_len', 'compound_score', 'punct%', 'positive_score', 'negative_score', 'neutral_score']
].reset_index(drop=True), pd.DataFrame(tfidf_train.toarray())], axis=1)

# Concatenating the features columns with the TF-IDF features of the testing set
X_test_vect = pd.concat([X_test[
    ['tweet_id', 'avg_word_length', 'sia_positive_word_rate', 'sia_negative_word_rate', 'neutral_score', 
     'stopword_count', 'body_len', 'compound_score', 'punct%', 'positive_score', 'negative_score', 'neutral_score']
].reset_index(drop=True), pd.DataFrame(tfidf_test.toarray())], axis=1)

# Displaying the head (first few rows) of the X_train_vect DataFrame
X_train_vect.head()

# Displaying the head (first few rows) of the X_train_vect DataFrame
X_test_vect.head()

### Save vectorized dataframe

X_train_vect.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\X_train_vect.csv', mode='w', index=False)

X_test_vect.to_csv('C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\X_test_vect.csv', mode='w', index=False)

### N-Grams

# Initialize a CountVectorizer object with ngram_range=(2,2) to create bigrams
vectorizer = CountVectorizer(ngram_range=(2, 2))

# Fit and transform the text column of the dataframe
bigrams = vectorizer.fit_transform(data_cleaned_features['cleaned_text'])

# Convert the result to a dataframe
df_bigrams = pd.DataFrame(bigrams.toarray(), columns=vectorizer.get_feature_names_out())

# Display the first few rows of the dataframe
df_bigrams.head()

###  Final features Dataframe

# First, reset the index of all dataframes to ensure they align correctly
X_train_vect.reset_index(drop=True, inplace=True)
df_bigrams.reset_index(drop=True, inplace=True)

# Then, concatenate all the dataframes along axis=1 (i.e., columns)
df_all_features = pd.concat([X_train_vect, df_bigrams], axis=1)

### Save final features dataset

data.to_csv("C:\\Users\\danij\\Documents\\UC3M\\TFG\\DATA\\tdidf_ngrams_features.csv", mode='w', index=False)
